requirements.txt
aiogram>=3.4.1,<4.0.0
httpx>=0.27.0,<1.0.0
feedparser>=6.0.11,<7.0.0
pydantic>=2.6.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
APScheduler>=3.10.4,<4.0.0
